backend/requirements.txt
fastapi==0.124.2
uvicorn==0.38.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.12.0
sentence-transformers>=2.6.0
torch>=2.0.0
umap-learn>=0.5.5
numpy>=1.21.0
huggingface-hub>=0.20.0
